requirements.txt
torch
TorchCRF
conllu
python-dotenv
morfessor
Morfessor-FlatCat
subword-nmt